stream-app/requirements.txt
fastapi
uvicorn
lightstreamer-client-lib 
httpx
requests
sqlalchemy
psycopg2-binary